fix bug in extract_names
Changed addr_book to address_book, the name of the parameter.

As I told you, it works, but it's because the variable is defined
outside of the function. And it dodgy. A function should be independent
when possible, not caring about the 'outside world'.

We'll see more about this a bit latter (scopes ;))